requirements.txt
openai>=1.0
pandas
requests
python-dotenv
joblib
fastapi
uvicorn[standard]